Split rustuna_js wasm outputs for node and web
