+++ rustuna_js/test/user_attr.mjs
@@ -1,7 +1,7 @@
 import test from "node:test";
 import assert from "node:assert";
 
-import * as rustuna from "../pkg/rustuna.js";
+import * as rustuna from "../pkg/node/rustuna.js";
 
 test("set_user_attr", () => {
 	const study = rustuna.create_study("test");
